What's new in Gerrit 2.8: Don't use `data-uri`
Using the `data-uri` option causes the images to be embedded in the
generated html as base64 data, but this does not seem to work very
well:

- Some images are not displayed at all on some browsers.  An example
  of this is the old change screen screenshot (before the image was
  rewritten by pngcrush) which did not display in Chrome on my Macbook
  Pro, but seemed OK in Chrome on Ubuntu.

- Some images are displayed with corruption.  For example the new
  change screen preferences screenshot is missing the bottom half and
  has strange artifacts across it, in Chrome on both Macbook Pro and
  Ubuntu.

After removing the `data-uri` option, the images are simply linked as
in a regular html webpage, and they are all rendered properly.

The disadvantage of this change is that the presentation can no longer
be distributed as a single html file.

Change-Id: I7f5409f83806579d78bde00b21702e0da49f3189