Update patch set 4
Patch Set 4:

> That's great, thanks a lot! May be we can host the result html on
 > gerrit-review or some other easy accessible place?

Would be nice having it on gerrit-review. We can discuss this on the Gerrit User Summit. This is where I want to use these slides to present the Gerrit 2.9 features. I wouldn't like to make a big annoucement of this before I had a chance to present it :)

Piotr Szwed once was setting up a public build for the training material [1] and the build results are available at [2]. Maybe it could be extended to also build the newly added presentations.

 [1] https://groups.google.com/d/msg/repo-discuss/rdmzByl3_M4/e55xqPa1skMJ
 [2] http://gerrit-training.scmforge.com/

Patch-set: 4