Fix screenshot with related changes
The green tilde icon was wrongly explained as 'indirect ancestor', but
actually it marks an 'indirect descendant'. Also update the
description of the black dot icon. It shows a merged commit and it is
not always the case that this commit bypassed code review, as claimed
by the old description.

Change-Id: I34808034dc3e1d70158b683b346e3a823920a06b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>