Update screenshots of new change screen that show the project links
There are two project links in the change info block on the change
screen, the name of the project and an icon on the right side of the
project name. The old screenshots showed a query icon, and the project
name linked to the project admin screen. This behaviour was changed
and the project name now links to the default dashboard of the project
and the query icon was replaced by a gear icon, to link to the project
admin screen.

A few other screenshots have been updated as well to be consistent
with the new screenshots.

Change-Id: I875b00a917e8fad37a7f7c4ba8b01f5a091c0996
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>